split Sender and Receiver into interface and implementation
