+++ src/main/java/st/extreme/klingklong/Receiver.java
@@ -1,51 +1,20 @@
 package st.extreme.klingklong;
 
-import static st.extreme.klingklong.util.Horn.honk;
-import static st.extreme.klingklong.util.Temperature.COSY;
-import static st.extreme.klingklong.util.Temperature.HOT;
+/**
+ * The interface for the receiver of the endpoint
+ */
+public interface Receiver {
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.ServerSocket;
-import java.net.Socket;
-import java.net.UnknownHostException;
-import java.util.concurrent.Semaphore;
-import java.util.function.Consumer;
+  /**
+   * Start the receiver
+   */
+  void start();
 
-public class Receiver extends Thread {
-  private final int listeningPort;
-  private final Consumer<String> messageConsumer;
-  private final Semaphore readSemaphore;
-
-  public Receiver(int listeningPort, Consumer<String> messageConsumer, Semaphore readSemaphore) throws UnknownHostException {
-    this.listeningPort = listeningPort;
-    this.messageConsumer = messageConsumer;
-    this.readSemaphore = readSemaphore;
-  }
-
-  @Override
-  public void run() {
-    try (ServerSocket serverSocket = new ServerSocket(listeningPort);
-        Socket listeningSocket = serverSocket.accept();
-        BufferedReader in = new BufferedReader(new InputStreamReader(listeningSocket.getInputStream()));) {
-      readSemaphore.acquireUninterruptibly();
-      String inputLine;
-      while ((inputLine = in.readLine()) != null) {
-        String message = Message.afterReceiving(inputLine);
-        if (Sender.STOP_SIGNAL.equals(message)) {
-          honk(HOT, "receiver got STOP signal");
-          break;
-        } else {
-          messageConsumer.accept(message);
-        }
-      }
-    } catch (IOException e) {
-      honk(COSY, "Exception caught when trying to listen on port " + listeningPort + " or listening for a connection");
-      e.printStackTrace();
-    }
-    readSemaphore.release();
-    honk(COSY, "receiver thread is terminating now");
-  }
+  /**
+   * Check if the receiver is still alive
+   * 
+   * @return {@code true} if the receiver is alive, {@code false} otherwise.
+   */
+  boolean isAlive();
 
 }
